requirements.txt
# requirements.txt with compatible versions
flask==2.2.3
werkzeug==2.2.3  # This is the version compatible with Flask 2.2.3
gunicorn==20.1.0
uvicorn==0.23.2
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
grpcio==1.70.0
h5py==3.13.0
idna==3.10
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
mypy==1.15.0
mypy-extensions==1.0.0
namex==0.0.8
numpy==2.0.2
opt-einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rich==13.9.4
ruff==0.9.9
setuptools==75.8.2
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
google-cloud-storage==2.16.0
gcsfs==2024.3.1
termcolor==2.5.0
tomli==2.2.1
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
wrapt==1.17.2
fastapi==0.115.0
python-multipart==0.0.5
